+++ app/charts/table/table-content.tsx
@@ -1,11 +1,10 @@
-import Flex from "@/components/flex";
-import { Box } from "@mui/material";
+import { Box, TableSortLabel } from "@mui/material";
 import * as React from "react";
 import { HeaderGroup } from "react-table";
-import { Observation } from "@/domain/data";
-import { Icon } from "@/icons";
-import { SORTING_ARROW_WIDTH } from "@/charts/table/constants";
-import { ColumnMeta } from "@/charts/table/table-state";
+import Flex from "../../components/flex";
+import { Observation } from "../../domain/data";
+import { SORTING_ARROW_WIDTH } from "./constants";
+import { ColumnMeta } from "./table-state";
 
 /** Workaround because react-window can't pass props to inner element */
 type TableContentProps = {
@@ -93,25 +92,12 @@ export const TableContent = ({ children }: { children: React.ReactNode }) => {
                             : "flex-start",
                       }}
                     >
-                      <Box>{column.render("Header")}</Box>
-                      {isCustomSorted && (
-                        <Box
-                          sx={{
-                            width: SORTING_ARROW_WIDTH,
-                            flexShrink: 0,
-                            mr: "-11px",
-                          }}
-                        >
-                          <Icon
-                            name={
-                              column.isSortedDesc
-                                ? "sortDescending"
-                                : "sortAscending"
-                            }
-                            size={SORTING_ARROW_WIDTH}
-                          />
-                        </Box>
-                      )}
+                      <TableSortLabel
+                        active={isCustomSorted}
+                        direction={column.isSortedDesc ? "desc" : "asc"}
+                      >
+                        <Box>{column.render("Header")}</Box>
+                      </TableSortLabel>
                     </Flex>
                   </Box>
                 );